RockPaperScissors: Added chunk of symmetric peer-to-peer TCP network connection.
A symmetric connection if some kind of peer-to-peer connection where no mediator server is needed for the peers to find each other. Instead, the peers hostname and port number is explicitly specified, and they are searching for each other on the net with trying to connect to and accept from the same place.
The current code chunk contains a basin session class which resolves the remote hostname and then calls an empty method chunk that will be filled with the sysmmetric connection establishment code.
The main program tries the whole thing out by resolving the remote host given as command line argument. Command line also requires the port number for "near future use".